范.修改:shopDao轉interface (#24)
